+++ protocol/daemons/shared/paginated_grpc_request.go
@@ -1,4 +1,4 @@
-package types
+package shared
 
 import (
 	"context"
